requirements.txt

streamlit
requests
feedparser
sentence-transformers
chromadb
python-dotenv
pandas
scikit-learn
tqdm
